02_Japan_Equity_Multi_Factor_Strategy/requirements.txt
requests
pandas
numpy
yfinance
aiohttp
japanize-matplotlib
matplotlib
seaborn
lxml
tqdm
nest_asyncio
xlrd
openpyxl
python-dotenv
